[frontend] refactoring some playbooks components
